// resources/js/api/generated/models/networkInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { NetworkInfoLabels } from './networkInfoLabels';

export interface NetworkInfo {
  created: string;
  driver: string;
  id: string;
  internal: boolean;
  labels: NetworkInfoLabels;
  name: string;
  scope: string;
  subnet: string;
  unused: boolean;
}